pyproject.toml
[tool.poetry]
name = "cryptohack"
version = "0.1.0"
description = ""
authors = ["Dawid Karpiński <dawid.karp@int.pl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sympy = "^1.13.3"
pycryptodome = "^3.21.0"
galois = "^0.4.2"
requests = "^2.32.3"
ecdsa = "^0.19.0"
primefac = "^2.0.12"
factordb-python = "^1.3.0"
owiener = "^1.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
